locker-management/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9